requirements.txt
langgraph==0.2.*
langchain-core==0.3.*
requests==2.*
readability-lxml==0.8.*
beautifulsoup4==4.*
lxml==5.*
pydantic==2.*
python-dotenv==1.*
markdownify==0.13.*
duckduckgo-search==6.*
langchain-groq==0.1.*
langsmith==0.1.*
ddgs==0.1.*
openai==0.27.*
markdownify==0.13.*
streamlit==1.*
pandas==2.*
python-dotenv==1.*
tavily==0.1.*
ddgs==0.1.*